requirements.txt
torch>=2.0.0
transformers>=4.30.0
openai>=1.0.0
pandas>=1.5.0
openpyxl>=3.1.0
Pillow>=10.0.0
requests>=2.31.0
tqdm>=4.65.0